# apps/core-server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/commerce-api/inventory-items/table": {
      "get": {
        "operationId": "InventoryItemsGatewayController_getTable",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/select": {
      "get": {
        "operationId": "InventoryItemsGatewayController_select",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter by label",
            "schema": {
              "example": "united",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "values",
            "required": false,
            "in": "query",
            "description": "Comma-separated values to fetch specific options",
            "schema": {
              "example": "1,2,3",
              "type": "string"
            }
          },
          {
            "name": "excludeIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated IDs to exclude from results (already selected)",
            "schema": {
              "example": "5,10",
              "type": "string"
            }
          },
          {
            "name": "valueKey",
            "required": false,
            "in": "query",
            "description": "Column name for option value",
            "schema": {
              "default": "id",
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "labelKey",
            "required": false,
            "in": "query",
            "description": "Column name for option label",
            "schema": {
              "default": "name",
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "descriptionKey",
            "required": false,
            "in": "query",
            "description": "Column name for option description",
            "schema": {
              "example": "description",
              "type": "string"
            }
          },
          {
            "name": "additionalKeys",
            "required": false,
            "in": "query",
            "description": "Comma-separated column names to include in option.additionals",
            "schema": {
              "example": "locationName,availableQuantity",
              "type": "string"
            }
          },
          {
            "name": "groupIdKey",
            "required": false,
            "in": "query",
            "description": "Column name for group ID",
            "schema": {
              "example": "regionId",
              "type": "string"
            }
          },
          {
            "name": "orderByKey",
            "required": false,
            "in": "query",
            "description": "Column name to sort options by",
            "schema": {
              "default": "name",
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "orderDirection",
            "required": false,
            "in": "query",
            "description": "Sort direction for options",
            "schema": {
              "default": "asc",
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "supplierId",
            "required": false,
            "in": "query",
            "description": "Optional supplier ID to filter linked inventory items",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items": {
      "post": {
        "operationId": "InventoryItemsGatewayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}": {
      "get": {
        "operationId": "InventoryItemsGatewayController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      },
      "patch": {
        "operationId": "InventoryItemsGatewayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      },
      "delete": {
        "operationId": "InventoryItemsGatewayController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/levels": {
      "get": {
        "operationId": "InventoryItemsGatewayController_findLevels",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/ledger": {
      "get": {
        "operationId": "InventoryItemsGatewayController_findLedger",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/levels/table": {
      "get": {
        "operationId": "InventoryItemsGatewayController_getLevelsTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/ledger/table": {
      "get": {
        "operationId": "InventoryItemsGatewayController_getLedgerTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/location-stock": {
      "get": {
        "operationId": "InventoryItemsGatewayController_getLocationStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/batches/table": {
      "get": {
        "operationId": "InventoryItemsGatewayController_getBatchesTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/storage-location-configs/table": {
      "get": {
        "operationId": "InventoryItemsGatewayController_getStorageLocationConfigsTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/storage-location-configs": {
      "post": {
        "operationId": "InventoryItemsGatewayController_createStorageLocationConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStorageLocationConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
    "/commerce-api/inventory-items/{id}/storage-location-configs/{configId}": {
      "patch": {
        "operationId": "InventoryItemsGatewayController_updateStorageLocationConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "configId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStorageLocationConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      },
      "delete": {
        "operationId": "InventoryItemsGatewayController_deleteStorageLocationConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "configId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Inventory Items"
        ]
      }
    },
...
    "/commerce-api/items/table": {
      "get": {
        "description": "Returns paginated, filtered, and sorted items using server-stored table state.",
        "operationId": "ItemsGatewayController_getItemsTable",
        "parameters": [
          {
            "name": "buId",
            "required": true,
            "in": "query",
            "description": "Business unit ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Items table retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemsTableResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get items table for a business unit",
        "tags": [
          "Commerce - Items"
        ]
      }
    },
...
    "/commerce-api/items/{id}": {
      "get": {
        "description": "Returns item with full details including options and variants.",
        "operationId": "ItemsGatewayController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Item not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an item by ID",
        "tags": [
          "Commerce - Items"
        ]
      },
      "patch": {
        "operationId": "ItemsGatewayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Item not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an item",
        "tags": [
          "Commerce - Items"
        ]
      },
      "delete": {
        "operationId": "ItemsGatewayController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Item not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an item",
        "tags": [
          "Commerce - Items"
        ]
      }
    },
...
    "/commerce-api/stock-adjustments/{id}/lines/{lineId}/items/table": {
      "get": {
        "operationId": "StockAdjustmentsGatewayController_getLineItemsTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Stock Adjustments"
        ]
      }
    },
...
    "/commerce-api/stock-adjustments/{id}/lines/{lineId}/items/{itemId}": {
      "patch": {
        "operationId": "StockAdjustmentsGatewayController_updateLineItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockAdjustmentLineItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Stock Adjustments"
        ]
      },
      "delete": {
        "operationId": "StockAdjustmentsGatewayController_removeLineItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Stock Adjustments"
        ]
      }
    },
...
    "/commerce-api/suppliers/items/{itemId}": {
      "delete": {
        "operationId": "SuppliersGatewayController_unlinkItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vritti API Nexus",
    "description": "Internal API for Vritti Nexus Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3001",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "CreateInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Item name",
            "example": "Basmati Rice"
          },
          "code": {
            "type": "string",
            "description": "Item code",
            "example": "RAW-RICE-BAS"
          },
          "type": {
            "type": "string",
            "description": "Item type",
            "enum": [
              "MATERIAL",
              "PRODUCT"
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "uomId": {
            "type": "string",
            "description": "Unit of measure ID"
          }
        },
        "required": [
          "name",
          "code",
          "type",
          "categoryId",
          "uomId"
        ]
      },
      "CreateStorageLocationConfigDto": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string",
            "description": "Storage location ID"
          },
          "reorderLevel": {
            "type": "number",
            "description": "Minimum stock level threshold for reorder at this location"
          }
        },
        "required": [
          "locationId",
          "reorderLevel"
        ]
      },
      "UpdateStorageLocationConfigDto": {
        "type": "object",
        "properties": {
          "reorderLevel": {
            "type": "number",
            "description": "Minimum stock level threshold for reorder at this location"
          }
        },
        "required": [
          "reorderLevel"
        ]
      },
      "UpdateInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Item name"
          },
          "code": {
            "type": "string",
            "description": "Item code"
          },
          "type": {
            "type": "string",
            "description": "Item type",
            "enum": [
              "MATERIAL",
              "PRODUCT"
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID"
          },
          "description": {
            "type": "object",
            "description": "Description"
          },
          "uomId": {
            "type": "string",
            "description": "Unit of measure ID"
          }
        }
      },
...
      "ItemsTableResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemResponseDto"
            }
          },
          "count": {
            "type": "number"
          },
          "state": {
            "type": "object",
            "description": "Current active filter/sort/visibility state"
          },
          "activeViewId": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "result",
          "count",
          "state"
        ]
      },
...
      "ItemDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Item ID"
          },
          "businessUnitId": {
            "type": "string",
            "description": "Business unit ID"
          },
          "categoryId": {
            "type": "object",
            "description": "Category ID",
            "nullable": true
          },
          "categoryName": {
            "type": "object",
            "description": "Category name",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Item type",
            "enum": [
              "PRODUCT",
              "SERVICE"
            ]
          },
          "code": {
            "type": "string",
            "description": "Item code"
          },
          "name": {
            "type": "string",
            "description": "Item name"
          },
          "description": {
            "type": "object",
            "description": "Item description",
            "nullable": true
          },
          "taxGroupId": {
            "type": "object",
            "description": "Tax group ID",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Whether the item is available"
          },
          "sortOrder": {
            "type": "number",
            "description": "Display sort order"
          },
          "options": {
            "description": "Item options",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemOptionResponseDto"
            }
          },
          "variants": {
            "description": "Item variants",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemVariantResponseDto"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp of creation"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp of last update"
          }
        },
        "required": [
          "id",
          "businessUnitId",
          "type",
          "code",
          "name",
          "isAvailable",
          "sortOrder",
          "options",
          "variants",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}